requirements.txt
Flask==2.3.2
python-docx==0.8.11
pdfkit==1.0.0
pypandoc==1.10
docxtpl==0.16.7
htmldocx==0.0.5
python-docx-replace==0.4.0
unoconv==0.9.0 